updated quantity to int as it should be a whole number.

+++ src/main/java/com/auction/product/dto/BuyNowProduct.java
@@ -6,6 +6,6 @@ public record BuyNowProduct(
         @NotNull(message = "Product Id mandatory")
         Long productId,
         @NotNull(message = "Quantity Id mandatory")
-        double quantity
+        int quantity
 ) {
 }
